Renames SchemaImporter to DocSchemaImporter
